Integrated libPAK fixes with pakGUI's hardcoded version of libPAK.
Fixed casting warning in pakEXE.

+++ FilePAK-Debug/main.cpp
@@ -185,7 +185,7 @@ string chooseEntry()
 	vector<string> entries = pak.getAllPAKEntries();
 	char choice;
 	int pages, page = 0;
-	pages = ceil((double) entries.size()/7);
+	pages = (int) ceil(entries.size()/7.0);
 
 	for(;;)
 	{
